move getPath, getHost and getToken utility functions into lib
use "redcap.uits.iu.edu" as fallback host if no other host has been
provided for unit testing

-> don't change unit test behaviour (even if setting a host is not
sensible default when provided for use by other institutions)
